+++ libs/arcadeshield/_locales/arcadeshield-strings.json
@@ -23,6 +23,7 @@
   "{id:category}Helpers": "Helpers",
   "{id:category}ScreenBitmap": "ScreenBitmap",
   "{id:category}Screenhelpers": "Screenhelpers",
+  "{id:category}__screen_internal": "__screen_internal",
   "{id:category}_helpers_workaround": "_helpers_workaround",
   "{id:group}Compare": "Compare",
   "{id:group}Create": "Create",
